extraced main game event handelr
