+++ handlers/history_viewer.py
@@ -1,8 +1,9 @@
 from typing import Optional
 import color
+from handlers.main_game_event_handler import MainGameEventHandler
 import keys
 from engine import Engine
-from handlers.input_handlers import EventHandler, MainGameEventHandler
+from handlers.input_handlers import EventHandler
 
 
 import tcod
