+++ handlers/look_handler.py
@@ -2,7 +2,7 @@ from typing import Optional
 from handlers.select_index_handler import SelectIndexHandler
 import keys
 from engine import Engine
-from handlers.input_handlers import MainGameEventHandler
+from handlers.main_game_event_handler import MainGameEventHandler
 from handlers.input_handlers import ActionOrHandler
