+++ handlers/pause_viewer.py
@@ -1,7 +1,8 @@
 import color
 import exceptions
+from handlers.main_game_event_handler import MainGameEventHandler
 import keys
-from handlers.input_handlers import EventHandler, MainGameEventHandler
+from handlers.input_handlers import EventHandler
 from typing import Optional
